Refactor: CourseCarousel component
Refactors `CourseCarousel.tsx` into smaller, more manageable files, including a hook for carousel logic, a component for navigation arrows, and a component for scroll indicators. This improves code organization and maintainability without altering existing functionality.
