+++ src/components/dashboard/CourseCarousel.tsx
@@ -1,12 +1,10 @@
 
-import React, { useState, useCallback, useEffect } from 'react';
-import { ChevronLeft, ChevronRight } from 'lucide-react';
-import { EmblaOptionsType } from 'embla-carousel';
-import useEmblaCarousel from 'embla-carousel-react';
-import { Button } from '@/components/ui/button';
+import React from 'react';
 import CourseCardWithProgress from './CourseCardWithProgress';
 import { Podcast } from '@/types';
-import { useIsMobile } from '@/hooks/use-mobile';
+import { useCourseCarousel } from './carousel/useCourseCarousel';
+import CourseCarouselNavigationArrows from './carousel/CourseCarouselNavigationArrows';
+import CourseCarouselScrollIndicators from './carousel/CourseCarouselScrollIndicators';
 
 interface CourseCarouselProps {
   title: string;
@@ -30,60 +28,16 @@ const CourseCarousel: React.FC<CourseCarouselProps> = ({
   onToggleSave,
   onCourseClick
 }) => {
-  const isMobile = useIsMobile();
-  const [canScrollPrev, setCanScrollPrev] = useState(false);
-  const [canScrollNext, setCanScrollNext] = useState(false);
-
-  // Unified Embla configuration optimized for mobile
-  const options: EmblaOptionsType = {
-    align: 'start',
-    slidesToScroll: 1,
-    containScroll: 'trimSnaps',
-    dragFree: true, // Better touch handling
-    skipSnaps: false,
-    startIndex: 0,
-  };
-
-  const [emblaRef, emblaApi] = useEmblaCarousel(options);
-
-  const scrollPrev = useCallback(() => {
-    if (emblaApi) {
-      console.log('CourseCarousel: Scrolling to previous slide');
-      emblaApi.scrollPrev();
-    }
-  }, [emblaApi]);
-
-  const scrollNext = useCallback(() => {
-    if (emblaApi) {
-      console.log('CourseCarousel: Scrolling to next slide');
-      emblaApi.scrollNext();
-    }
-  }, [emblaApi]);
-
-  const onSelect = useCallback(() => {
-    if (!emblaApi) return;
-    setCanScrollPrev(emblaApi.canScrollPrev());
-    setCanScrollNext(emblaApi.canScrollNext());
-    console.log('CourseCarousel: Carousel state updated', {
-      canScrollPrev: emblaApi.canScrollPrev(),
-      canScrollNext: emblaApi.canScrollNext(),
-      selectedIndex: emblaApi.selectedScrollSnap(),
-      totalSlides: courses.length
-    });
-  }, [emblaApi, courses.length]);
-
-  useEffect(() => {
-    if (!emblaApi) return;
-    console.log('CourseCarousel: Embla API initialized with', courses.length, 'courses');
-    onSelect();
-    emblaApi.on('select', onSelect);
-    emblaApi.on('reInit', onSelect);
-
-    return () => {
-      emblaApi.off('select', onSelect);
-      emblaApi.off('reInit', onSelect);
-    };
-  }, [emblaApi, onSelect, courses.length]);
+  const {
+    emblaRef,
+    emblaApi,
+    canScrollPrev,
+    canScrollNext,
+    selectedIndex,
+    scrollPrev,
+    scrollNext,
+    isMobile
+  } = useCourseCarousel({ coursesLength: courses.length });
 
   if (courses.length === 0) {
     return (
@@ -138,77 +92,22 @@ const CourseCarousel: React.FC<CourseCarouselProps> = ({
             </div>
           </div>
 
-          {/* Mobile navigation arrows - always show both if more than 1 course */}
-          {isMobile && courses.length > 1 && (
-            <div className="absolute inset-0 pointer-events-none">
-              {/* Left arrow */}
-              <Button
-                variant="outline"
-                size="icon"
-                className="absolute left-6 top-1/2 -translate-y-1/2 bg-white/90 backdrop-blur-sm shadow-lg z-20 h-8 w-8 border-gray-200 pointer-events-auto"
-                onClick={scrollPrev}
-                disabled={!canScrollPrev}
-                style={{ opacity: canScrollPrev ? 1 : 0.3 }}
-              >
-                <ChevronLeft className="h-4 w-4" />
-              </Button>
-
-              {/* Right arrow */}
-              <Button
-                variant="outline"
-                size="icon"
-                className="absolute right-6 top-1/2 -translate-y-1/2 bg-white/90 backdrop-blur-sm shadow-lg z-20 h-8 w-8 border-gray-200 pointer-events-auto"
-                onClick={scrollNext}
-                disabled={!canScrollNext}
-                style={{ opacity: canScrollNext ? 1 : 0.3 }}
-              >
-                <ChevronRight className="h-4 w-4" />
-              </Button>
-            </div>
-          )}
-
-          {/* Desktop navigation arrows - only show when more than 4 courses */}
-          {!isMobile && courses.length > 4 && (
-            <>
-              <Button
-                variant="outline"
-                size="icon"
-                className="absolute left-0 top-1/2 -translate-y-1/2 -translate-x-4 bg-white shadow-lg z-10 h-10 w-10"
-                onClick={scrollPrev}
-                disabled={!canScrollPrev}
-              >
-                <ChevronLeft className="h-4 w-4" />
-              </Button>
-
-              <Button
-                variant="outline"
-                size="icon"
-                className="absolute right-0 top-1/2 -translate-y-1/2 translate-x-4 bg-white shadow-lg z-10 h-10 w-10"
-                onClick={scrollNext}
-                disabled={!canScrollNext}
-              >
-                <ChevronRight className="h-4 w-4" />
-              </Button>
-            </>
-          )}
+          <CourseCarouselNavigationArrows
+            isMobile={isMobile}
+            coursesLength={courses.length}
+            canScrollPrev={canScrollPrev}
+            canScrollNext={canScrollNext}
+            scrollPrev={scrollPrev}
+            scrollNext={scrollNext}
+          />
         </div>
 
-        {/* Mobile scroll indicator dots */}
-        {isMobile && courses.length > 1 && (
-          <div className="flex justify-center mt-4 space-x-2">
-            {courses.map((_, index) => (
-              <button
-                key={index}
-                className={`w-2 h-2 rounded-full transition-colors ${
-                  index === (emblaApi?.selectedScrollSnap() || 0) 
-                    ? 'bg-miyo-800' 
-                    : 'bg-gray-300'
-                }`}
-                onClick={() => emblaApi?.scrollTo(index)}
-              />
-            ))}
-          </div>
-        )}
+        <CourseCarouselScrollIndicators
+          isMobile={isMobile}
+          coursesLength={courses.length}
+          selectedIndex={selectedIndex}
+          emblaApi={emblaApi}
+        />
       </div>
     </div>
   );
